test: split hosted delegation credential tests under line limit
Move Phase-1 focused rendered-stack and agent-auth coverage into owned
part_002/part_005 so test_hosted_delegation_payloads.py stays under 1500 lines.
